move event actions into own file
