+++ src/index.js
@@ -5,7 +5,7 @@ import { Provider } from "react-redux";
 import thunk from "redux-thunk";
 import { composeWithDevTools } from "redux-devtools-extension";
 import reducers from "./reducers";
-import { getEvents } from "./actions";
+import { getEvents } from "./actions/events";
 import App from "./App";
 
 const store = createStore(
