+++ src/screens/EventsScreen/index.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import type { Connector } from "react-redux";
 import type { NavigationScreenProp, NavigationState } from "react-navigation";
 import type { Event } from "../../integrations/cms";
-import { updateEvents } from "../../actions";
+import { updateEvents } from "../../actions/events";
 import Component from "./component";
 
 type OwnProps = {
